# distsql: add external storage to the hashJoiner