This PR is composed of two commits. The first one creates an interface and puts the current in-memory implementation behind it and the second one implements an on-disk version of this interface. The in-memory changes result have the following impact on performance:
```
$ make bench PKG=./pkg/sql/distsqlrun BENCHES=BenchmarkHashJoiner TESTFLAGS="-count 20" > hashchanges
$ benchstat hashbench hashchanges
name                          old time/op  new time/op  delta
HashJoiner/InputSize=0-8      9.53µs ± 7%  4.44µs ± 1%  -53.44%  (p=0.000 n=20+18)
HashJoiner/InputSize=4-8      18.4µs ± 2%  14.4µs ± 2%  -21.58%  (p=0.000 n=18+20)
HashJoiner/InputSize=16-8     38.8µs ± 2%  40.2µs ± 6%   +3.69%  (p=0.000 n=19+20)
HashJoiner/InputSize=256-8     466µs ± 4%   582µs ± 5%  +24.84%  (p=0.000 n=20+20)
HashJoiner/InputSize=4096-8   7.83ms ± 2%  9.41ms ± 3%  +20.07%  (p=0.000 n=20+20)
HashJoiner/InputSize=65536-8   133ms ± 6%   156ms ± 4%  +16.81%  (p=0.000 n=20+20)
```
I believe this can be improved with changes to the marking API.

The on-disk implementation is currently tested by setting `testingKnobMemLimit = 1` to force immediate use of disk.

There is still some work to be done regarding:
1) Changing tests to test both the in-memory and on-disk implementation. The interaction between these two also needs to be tested.
2) The marking API has to be changed (this is not necessary for functionality but if there is time it's worth it to tackle this second point)

@arjunravinarayan will be testing this on a production cluster in the meantime

cc @cockroachdb/distsql