@arjunravinarayan 2017-07-24
PSA that production testing is running behind schedule due to an Azure `fuseblk` issue, so we may not have empirical results in a timely fashion.

---







*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/17192#-:-KppVwg6j8SEsUi78YZB:buc3m5n)*
<!-- Sent from Reviewable.io -->