@RaduBerinde 2017-07-24
A suggestion to improve test coverage: add a new logic test config `5node-low-mem` that limits the working set to something small (see `logicTestConfigs`); and add this config to the `distsql_*` testfiles.