@andreimatei 2017-07-24
Some comments on the interface for now.

---

Review status: 0 of 5 files reviewed at latest revision, 8 unresolved discussions, some commit checks failed.

---

*[pkg/sql/distsqlrun/hash_row_container.go, line 31 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17192#-Kpqf6FdK-BV0m_jhR7U:-Kpqf6FdK-BV0m_jhR7V:bl3nxv4) ([raw file](https://github.com/cockroachdb/cockroach/blob/efba82c78b4c0eabee0e6823cebb7e02fccab8e0/pkg/sql/distsqlrun/hash_row_container.go#L31)):*
> ```Go
> 
> // hashRowContainer is a container used to store rows according to an encoding
> // of given equality columns. The stored rows can then be probed to return a
> ```

which "given equality columns"? They are not given to this interface. I'd make `AddRow,SetMarks,NewBucketIterator` take a `colMap` specifying which columns from the passed-in row should be used to construct the key and in which order.

---

*[pkg/sql/distsqlrun/hash_row_container.go, line 43 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17192#-KpqjVnlSfmMr2pK-YOK:-KpqjVnlSfmMr2pK-YOL:bcbqekw) ([raw file](https://github.com/cockroachdb/cockroach/blob/efba82c78b4c0eabee0e6823cebb7e02fccab8e0/pkg/sql/distsqlrun/hash_row_container.go#L43)):*
> ```Go
> 	// TODO(asubiotto): Switch this to mark through the iterator. Will avoid the
> 	// caller having to allocate marks.
> 	SetMarks(context.Context, sqlbase.EncDatumRow, []bool) error
> ```

I was very confused when I read this. The method comment doesn't suggest the last argument (which, btw, should be named, otherwise it's even more mysterious, and it also seems like you're referring to it by name in the comment). You need to explain the argument and how it needs to match the array of indexes in a bucked.
To keep this API, you'll have to explain the use - I guess something about `ON` clauses and how they can have other expressions than the equality columns.

It seems to me that this doesn't belong in this interface, which is just a hashtable. Can it be extracted and the markings be maintained by the client?
Or if there's big benefits to holding the markings here, I think we should do what you say - do the marking through an interator.

---

*[pkg/sql/distsqlrun/hash_row_container.go, line 50 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17192#-KpqgjPs6aHQpaiQFPN6:-KpqgjPtM7zsa4_ZrCHS:b-ff87gm) ([raw file](https://github.com/cockroachdb/cockroach/blob/efba82c78b4c0eabee0e6823cebb7e02fccab8e0/pkg/sql/distsqlrun/hash_row_container.go#L50)):*
> ```Go
> 
> 	// NewIterator returns a rowIterator that iterates over unmarked rows.
> 	NewIterator(context.Context) rowIterator
> ```

`IterateUnmarked()` ?

---

*[pkg/sql/distsqlrun/hash_row_container.go, line 97 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17192#-Kpqd2nO0WjiHY2s_wW8:-Kpqd2nO0WjiHY2s_wW9:b-8eyukp) ([raw file](https://github.com/cockroachdb/cockroach/blob/efba82c78b4c0eabee0e6823cebb7e02fccab8e0/pkg/sql/distsqlrun/hash_row_container.go#L97)):*
> ```Go
> 	// TODO(asubiotto): This memRowContainer then has an underlying
> 	// sqlbase.RowContainer. This can be cleaned up.
> 	*memRowContainer
> ```

seems weird to me that `buckets` stores indexes into this guy. Can't we have one container per bucket? Or not use containers at all?

Also, are you sure you want to embed the container? What methods are you inheriting?

---

*[pkg/sql/distsqlrun/hash_row_container.go, line 107 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17192#-KpqcjM2x6uL4sLhDdt5:-KpqcjM2x6uL4sLhDdt6:b-vq4rmy) ([raw file](https://github.com/cockroachdb/cockroach/blob/efba82c78b4c0eabee0e6823cebb7e02fccab8e0/pkg/sql/distsqlrun/hash_row_container.go#L107)):*
> ```Go
> 	// buckets contains the indices into memRowContainer for a given group
> 	// key (which is the encoding of storedEqCols).
> 	buckets map[string][]int
> ```

`[]byte` can't be a map key? :(

---

*[pkg/sql/distsqlrun/hash_row_container.go, line 114 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17192#-Kpqe3dcZDA9K-Y5nx74:-Kpqe3dcZDA9K-Y5nx75:b8m2svk) ([raw file](https://github.com/cockroachdb/cockroach/blob/efba82c78b4c0eabee0e6823cebb7e02fccab8e0/pkg/sql/distsqlrun/hash_row_container.go#L114)):*
> ```Go
> 	// {stored, probe}eqCols contain the indices of the columns of a row that
> 	// are encoded and used as a key into buckets. probeEqCols are used in
> 	// NewBucketIterator() and SetMarks(), and storedEqCols are used in
> ```

I find this very confusing :)
I think you should either explain how this is going to be used - probably has something to do with a join having two input tables with the columns ordered differently, or, better yet in my opinion, make all operations that need a set of column indexes take a `colMap` - map from column index in the hash key to the column index in the encoded row.

---

*[pkg/sql/distsqlrun/hash_row_container.go, line 194 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17192#-Kpqc4PCKXbYhc8mwTQx:-Kpqc4PDhuyz5dZ1IJ61:b-2iv8ut) ([raw file](https://github.com/cockroachdb/cockroach/blob/efba82c78b4c0eabee0e6823cebb7e02fccab8e0/pkg/sql/distsqlrun/hash_row_container.go#L194)):*
> ```Go
> // addRowToBuckets is a helper function that encodes the equality columns of the
> // given row and appends the rowIdx to the matching bucket.
> func (h *hashMemRowContainer) addRowToBuckets(
> ```

I'd name this `addRowToBucket`, otherwise I was confused. You're adding a row to a single bucket. The name of the `buckets` member is less important, IMHO.

---

*[pkg/sql/distsqlrun/hashjoiner.go, line 143 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17192#-KpqiHatECTavr2IRwld:-KpqiHatECTavr2IRwle:b-doby3g) ([raw file](https://github.com/cockroachdb/cockroach/blob/efba82c78b4c0eabee0e6823cebb7e02fccab8e0/pkg/sql/distsqlrun/hashjoiner.go#L143)):*
> ```Go
> 		// draining will be swallowed, and the original error will be forwarded
> 		// to the consumer.
> 		log.Infof(ctx, "error making hash row container %s", err)
> ```

```
err = errors.Wrap(err, "error creating hash row container")
```

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/17192#-:-KpqkYLbh8Fsr7d5TUPv:b-5cz3r3)*
<!-- Sent from Reviewable.io -->