@asubiotto 2017-07-26
Completed the pending changes and the benchmark is looking better:
```
name                          old time/op  new time/op  delta
HashJoiner/InputSize=0-8      9.53µs ± 7%  4.51µs ± 3%  -52.71%  (p=0.000 n=20+20)
HashJoiner/InputSize=4-8      18.4µs ± 2%  13.6µs ± 2%  -25.67%  (p=0.000 n=18+19)
HashJoiner/InputSize=16-8     38.8µs ± 2%  36.0µs ± 1%   -7.32%  (p=0.000 n=19+18)
HashJoiner/InputSize=256-8     466µs ± 4%   516µs ± 2%  +10.83%  (p=0.000 n=20+20)
HashJoiner/InputSize=4096-8   7.83ms ± 2%  8.37ms ± 1%   +6.79%  (p=0.000 n=20+20)
HashJoiner/InputSize=65536-8   133ms ± 6%   137ms ± 3%   +2.83%  (p=0.001 n=20+20)
```
PTAL.

---

Review status: 0 of 5 files reviewed at latest revision, 8 unresolved discussions, some commit checks failed.

---

*[pkg/sql/distsqlrun/hash_row_container.go, line 31 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17192#-Kpqf6FdK-BV0m_jhR7U:-KpzUGtEwSuHTbHqnJPn:bsguhlu) ([raw file](https://github.com/cockroachdb/cockroach/blob/efba82c78b4c0eabee0e6823cebb7e02fccab8e0/pkg/sql/distsqlrun/hash_row_container.go#L31)):*
<details><summary><i>Previously, andreimatei (Andrei Matei) wrote…</i></summary><blockquote>

which "given equality columns"? They are not given to this interface. I'd make `AddRow,SetMarks,NewBucketIterator` take a `colMap` specifying which columns from the passed-in row should be used to construct the key and in which order.
</blockquote></details>

Answered this comment below.

---

*[pkg/sql/distsqlrun/hash_row_container.go, line 43 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17192#-KpqjVnlSfmMr2pK-YOK:-Kpv5QWfwawTx5637lZX:b-sg33vi) ([raw file](https://github.com/cockroachdb/cockroach/blob/efba82c78b4c0eabee0e6823cebb7e02fccab8e0/pkg/sql/distsqlrun/hash_row_container.go#L43)):*
<details><summary><i>Previously, andreimatei (Andrei Matei) wrote…</i></summary><blockquote>

I was very confused when I read this. The method comment doesn't suggest the last argument (which, btw, should be named, otherwise it's even more mysterious, and it also seems like you're referring to it by name in the comment). You need to explain the argument and how it needs to match the array of indexes in a bucked.
To keep this API, you'll have to explain the use - I guess something about `ON` clauses and how they can have other expressions than the equality columns.

It seems to me that this doesn't belong in this interface, which is just a hashtable. Can it be extracted and the markings be maintained by the client?
Or if there's big benefits to holding the markings here, I think we should do what you say - do the marking through an interator.
</blockquote></details>

I changed the marking API to be done through the iterator returned by `NewBucketIterator`. Unfortunately this cannot be extracted because the markings grow with the number of rows, so they are a memory liability. We could keep track of the markings from the `hashJoiner` in a data structure that also falls back to disk independently. However, I think that that would be quite messy.
I also added an explanation including the `ON` clauses.

---

*[pkg/sql/distsqlrun/hash_row_container.go, line 50 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17192#-KpqgjPs6aHQpaiQFPN6:-Kpv6KVUe5gYtNS56t6D:b3cgcyp) ([raw file](https://github.com/cockroachdb/cockroach/blob/efba82c78b4c0eabee0e6823cebb7e02fccab8e0/pkg/sql/distsqlrun/hash_row_container.go#L50)):*
<details><summary><i>Previously, andreimatei (Andrei Matei) wrote…</i></summary><blockquote>

`IterateUnmarked()` ?
</blockquote></details>

Changed to `NewUnmarkedIterator` to keep the style consistent.

---

*[pkg/sql/distsqlrun/hash_row_container.go, line 97 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17192#-Kpqd2nO0WjiHY2s_wW8:-Kpv6as7rCymwOgVFW0X:bpvf84q) ([raw file](https://github.com/cockroachdb/cockroach/blob/efba82c78b4c0eabee0e6823cebb7e02fccab8e0/pkg/sql/distsqlrun/hash_row_container.go#L97)):*
<details><summary><i>Previously, andreimatei (Andrei Matei) wrote…</i></summary><blockquote>

seems weird to me that `buckets` stores indexes into this guy. Can't we have one container per bucket? Or not use containers at all?

Also, are you sure you want to embed the container? What methods are you inheriting?
</blockquote></details>

We could have one container per bucket but that seems like unnecessary overhead. We could probably use no containers but that would require us to reuse the allocation optimization code in `sqlbase.rowContainer`. However, I think that this is out of the scope of this PR.

I'm not sure I do, I did it for convenience (using`Len` and `EncRow`). What's the cost of doing so?

---

*[pkg/sql/distsqlrun/hash_row_container.go, line 107 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17192#-KpqcjM2x6uL4sLhDdt5:-KpzU5gcNWf1DXrqE14h:b271wbi) ([raw file](https://github.com/cockroachdb/cockroach/blob/efba82c78b4c0eabee0e6823cebb7e02fccab8e0/pkg/sql/distsqlrun/hash_row_container.go#L107)):*
<details><summary><i>Previously, andreimatei (Andrei Matei) wrote…</i></summary><blockquote>

`[]byte` can't be a map key? :(
</blockquote></details>

Unfortunately not :(

---

*[pkg/sql/distsqlrun/hash_row_container.go, line 114 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17192#-Kpqe3dcZDA9K-Y5nx74:-KpzW17WtpEYmbSVvDeE:b3xsmey) ([raw file](https://github.com/cockroachdb/cockroach/blob/efba82c78b4c0eabee0e6823cebb7e02fccab8e0/pkg/sql/distsqlrun/hash_row_container.go#L114)):*
<details><summary><i>Previously, andreimatei (Andrei Matei) wrote…</i></summary><blockquote>

I find this very confusing :)
I think you should either explain how this is going to be used - probably has something to do with a join having two input tables with the columns ordered differently, or, better yet in my opinion, make all operations that need a set of column indexes take a `colMap` - map from column index in the hash key to the column index in the encoded row.
</blockquote></details>

The issue of using a `colMap` for every call is that:
1) The use of this `colMap` isn't expected to change throughout the lifetime of the hashRowContainer.
2) More importantly, the underlying diskRowContainer cannot handle different `colMap`s without some reworking and we would have to review the diskRowContainer because its sorting functionality would be undefined. This could be done but I don't think it's going to make it in as part of this change.
We talked yesterday about having some sort of reader/writer API in addition that would take a `colMap` as an initialization argument. This suffers from the same problems as above unless we restrict to one call maximum per lifetime of a rowContainer instance. This seems like a lot more surprising than not understanding the reference to equality columns in this interface.

What I have done to reference the equality columns is introduce an "Init" method to the API that takes in probe and stored columns to make it part of the interface. I think this is allows us to refer to the equality columns in the interface without introducing side-effects that would be difficult to manage.

---

*[pkg/sql/distsqlrun/hash_row_container.go, line 194 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17192#-Kpqc4PCKXbYhc8mwTQx:-KpzQDInAFHdkPpVNmVz:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/efba82c78b4c0eabee0e6823cebb7e02fccab8e0/pkg/sql/distsqlrun/hash_row_container.go#L194)):*
<details><summary><i>Previously, andreimatei (Andrei Matei) wrote…</i></summary><blockquote>

I'd name this `addRowToBucket`, otherwise I was confused. You're adding a row to a single bucket. The name of the `buckets` member is less important, IMHO.
</blockquote></details>

Done.

---

*[pkg/sql/distsqlrun/hashjoiner.go, line 143 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/17192#-KpqiHatECTavr2IRwld:-Kpzuwo9PB9DJWWfW6JX:b-896fix) ([raw file](https://github.com/cockroachdb/cockroach/blob/efba82c78b4c0eabee0e6823cebb7e02fccab8e0/pkg/sql/distsqlrun/hashjoiner.go#L143)):*
<details><summary><i>Previously, andreimatei (Andrei Matei) wrote…</i></summary><blockquote>

```
err = errors.Wrap(err, "error creating hash row container")
```
</blockquote></details>

Done.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/17192#-:-Kq-LLqC_c94kzPz0buA:bz7ei0m)*
<!-- Sent from Reviewable.io -->